1.删除无用代码；2.修改增加货物bug；3.增加库存查询（与之前库存查询不同，之前库存查询改为数量变动记录）；4.增加修改用户资料；5.完善邮箱通知功能；6.增加定时清理验证码图片代码；6.完善库存部分出库和全部出库，标准为数量；7.美化部分页面；8.修改出库货物bug

+++ blueprints/inventory.py
@@ -2,7 +2,7 @@ from flask import Blueprint, render_template, redirect, g, url_for, request, fla
 
 from config.decorators import login_required
 from config.exts import db
-from config.models import InboundModel, OutboundModel, InventoryModel, ConfigModel
+from config.models import InboundModel, OutboundModel, InventoryModel, ConfigModel, StockModel
 from utils.captchas import getCaptchaPic
 
 bp = Blueprint('inventory', __name__, url_prefix='/')
@@ -19,9 +19,16 @@ def inbound_manage():
 def inbound_manage_data():
     page = int(request.args.get('page', 1))
     limit = int(request.args.get('limit', 50))
-    data = InboundModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
-    total = data.total
-    data = data.items
+    key = request.args.get('key', '')
+    if key:
+        data = InboundModel.query.filter(InboundModel.rfid.contains(key)).order_by(
+            db.text('-create_date'))
+        total = data.count()
+    else:
+        data = InboundModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
+        total = data.total
+        data = data.items
+
     data_list = []
     for i in data:
         dit = {'id': i.id
@@ -43,6 +50,9 @@ def inbound_manage_data():
 def inbound_del():
     id_ = request.args.get('id')
     notice = InboundModel.query.filter_by(id=id_).first()
+    stock = StockModel.query.filter_by(rfid=notice.rfid).first()
+    if notice.num!=stock.num:
+        return jsonify({'code': 400, 'msg': '该货物已全部或部分出库，不允许修改！'})
     if not notice:
         return jsonify({'code': 400, 'msg': 'FAIL'})
     db.session.delete(notice)
@@ -71,7 +81,23 @@ def change_inbound_data():
     diameter = request.form.get('diameter')
     kg = request.form.get('kg')
     warehouse = request.form.get('warehouse')
+    if not num.isdigit() or not kg.isdigit():
+        return jsonify({'code': 400, 'msg': '数量和重量必须为数字！'})
     user = InboundModel.query.filter_by(id=id_).first()
+    if user.rfid != rfid:
+        if not check_rfid(rfid):
+            return jsonify({'code': 400, 'msg': 'RFID已存在！'})
+    stock = StockModel.query.filter_by(rfid=rfid).first()
+    if user.num!=stock.num:
+        return jsonify({'code': 400, 'msg': '该货物已全部或部分出库，不允许修改！'})
+    stock.state = state
+    stock.rfid = rfid
+    stock.num = num
+    stock.batch = batch
+    stock.lm = lm
+    stock.diameter = diameter
+    stock.kg = kg
+
     user.rfid = rfid
     user.num = num
     user.batch = batch
@@ -101,7 +127,15 @@ def add_inbound_data():
     diameter = request.form.get('diameter')
     kg = request.form.get('kg')
     warehouse = request.form.get('warehouse')
-    user = InboundModel(rfid=rfid, num=num, batch=batch, state=state, lm=lm, diameter=diameter, kg=kg, warehouse=warehouse)
+    if not check_rfid(rfid):
+        return jsonify({'code': 400, 'msg': 'RFID已存在！'})
+    if not num.isdigit() or not kg.isdigit():
+        return jsonify({'code': 400, 'msg': '数量和重量必须为数字！'})
+    user = InboundModel(rfid=rfid, num=num, batch=batch, state=state, lm=lm, diameter=diameter, kg=kg,
+                        warehouse=warehouse)
+    stock = StockModel(rfid=rfid, num=num, batch=batch, state=state, lm=lm, diameter=diameter, kg=kg,
+                        warehouse=warehouse)
+    db.session.add(stock)
     db.session.add(user)
     db.session.commit()
     return jsonify({'code': 200, 'msg': 'SUCCESS'})
@@ -118,9 +152,16 @@ def outbound_manage():
 def outbound_manage_data():
     page = int(request.args.get('page', 1))
     limit = int(request.args.get('limit', 50))
-    data = OutboundModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
-    total = data.total
-    data = data.items
+    key = request.args.get('key', '')
+    if key:
+        data = OutboundModel.query.filter(OutboundModel.rfid.contains(key)).order_by(
+            db.text('-create_date'))
+        total = data.count()
+    else:
+        data = OutboundModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
+        total = data.total
+        data = data.items
+
     data_list = []
     for i in data:
         dit = {'id': i.id
@@ -130,6 +171,7 @@ def outbound_manage_data():
             , 'diameter': i.diameter
             , 'kg': i.kg
             , 'batch': i.batch
+            , 'state': i.state
             , 'create_date': str(i.create_date)}
         data_list.append(dit)
     dic = {'code': 0, 'msg': 'SUCCESS', 'count': total, 'data': data_list}
@@ -153,22 +195,47 @@ def outbound_del():
 def change_outbound(id):
     user = OutboundModel.query.filter_by(id=id).first()
     data = {'rfid': user.rfid, 'num': user.num, 'batch': user.batch, 'id': user.id, 'lm': user.lm,
-            'diameter': user.diameter, 'kg': user.kg}
+            'diameter': user.diameter, 'kg': user.kg, 'state': user.state}
     return render_template('change_outbound.html', data=data)
+
+
+@bp.route('/restocking', methods=['GET', 'POST'])
+@login_required
+def restocking():
+    rfid = request.args.get('rfid')
+    outb = OutboundModel.query.filter_by(rfid=rfid).first()
+    outb.state = 3
+    inb = InboundModel.query.filter_by(rfid=rfid).first()
+    inb.state = 4
+    stock = StockModel.query.filter_by(rfid=rfid).first()
+    stock.num += int(outb.num)
+    stock.kg += int(outb.kg)
+    db.session.commit()
+    return jsonify({'code': 200, 'msg': 'SUCCESS'})
 
 
 @bp.route('/change_outbound_data', methods=['GET', 'POST'])
 @login_required
 def change_outbound_data():
     id_ = request.form.get('id')
-    rfid = request.form.get('rfid')
     num = request.form.get('num')
     batch = request.form.get('batch')
     lm = request.form.get('lm')
     diameter = request.form.get('diameter')
     kg = request.form.get('kg')
+    state = request.form.get('state')
+    if not num.isdigit() or not kg.isdigit():
+        return jsonify({'code': 400, 'msg': '数量和重量必须为数字！'})
     user = OutboundModel.query.filter_by(id=id_).first()
-    user.rfid = rfid
+    stock = StockModel.query.filter_by(rfid=user.rfid).first()
+    stock.state = state
+    stock.num = int(user.num)+int(stock.num)-int(num)
+    stock.batch = batch
+    stock.lm = lm
+    stock.diameter = diameter
+    stock.kg = int(user.kg)+int(stock.kg)-int(kg)
+
+    user.state = state
     user.num = num
     user.batch = batch
     user.lm = lm
@@ -188,12 +255,29 @@ def add_outbound_html():
 @login_required
 def add_outbound_data():
     rfid = request.form.get('rfid').replace(" ", "")
+    stock = StockModel.query.filter_by(rfid=rfid).first()
+    if not stock:
+        return jsonify({'code': 400, 'msg': 'RFID不存在'})
     num = request.form.get('num')
     batch = request.form.get('batch')
     lm = request.form.get('lm')
     diameter = request.form.get('diameter')
     kg = request.form.get('kg')
-    user = OutboundModel(rfid=rfid, num=num, batch=batch, lm=lm, diameter=diameter, kg=kg)
+    if not num.isdigit() or not kg.isdigit():
+        return jsonify({'code': 400, 'msg': '数量和重量必须为数字！'})
+    if int(stock.num) < int(num) or int(stock.kg) < int(kg):
+        return jsonify({'code': 400, 'msg': '库存不足'})
+    if int(stock.num) > int(num):
+        state = 2
+        stock.state = 5
+        stock.num = int(stock.num) - int(num)
+        stock.kg = int(stock.kg) - int(kg)
+    else:
+        state = 1
+        stock.state = 6
+        stock.num = 0
+        stock.kg = 0
+    user = OutboundModel(rfid=rfid, num=num, batch=batch, lm=lm, diameter=diameter, kg=kg, state=state)
     db.session.add(user)
     db.session.commit()
     return jsonify({'code': 200, 'msg': 'SUCCESS'})
@@ -202,7 +286,19 @@ def add_outbound_data():
 @bp.route('/inventory_check', methods=['GET', 'POST'])
 @login_required
 def inventory_check():
+    num = Calculate_inventory()
+    Inventory = InventoryModel(num=num)
+    db.session.add(Inventory)
+    db.session.commit()
     return render_template('inventory_check.html')
+
+
+def Calculate_inventory():
+    stock = StockModel.query.all()
+    num = 0
+    for i in stock:
+        num += int(i.num)
+    return num
 
 
 @bp.route('/inventory_check_data', methods=['GET', 'POST'])
@@ -210,9 +306,15 @@ def inventory_check():
 def inventory_check_data():
     page = int(request.args.get('page', 1))
     limit = int(request.args.get('limit', 50))
-    data = InventoryModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
-    total = data.total
-    data = data.items
+    key = request.args.get('key', '')
+    if key:
+        data = InventoryModel.query.filter(InventoryModel.rfid.contains(key)).order_by(
+            db.text('-create_date'))
+        total = data.count()
+    else:
+        data = InventoryModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
+        total = data.total
+        data = data.items
     data_list = []
     for i in data:
         dit = {'id': i.id
@@ -223,31 +325,10 @@ def inventory_check_data():
     return dic
 
 
-@bp.route('/notice_setting', methods=['GET', 'POST'])
-@login_required
-def notice_setting():
-    user = ConfigModel.query.filter_by(id=1).first()
-    data = {'Inventory_notification_thresholds': user.Inventory_notification_thresholds,
-            'notification': user.notification}
-    return render_template('notice_setting.html', data=data)
-
-
-@bp.route('/change_notification_data', methods=['GET', 'POST'])
-@login_required
-def change_notification_data():
-    thresholds = request.form.get('thresholds')
-    notification = request.form.get('notification')
-    user = ConfigModel.query.filter_by(id=1).first()
-    user.Inventory_notification_thresholds = thresholds
-    user.notification = notification
-    db.session.commit()
-    return jsonify({'code': 200, 'msg': 'SUCCESS'})
-
-
 @bp.route('/location_manage', methods=['GET', 'POST'])
 @login_required
 def location_manage():
-    return render_template('location_manage.html')
+    return render_template('location_manage.html', total=Calculate_inventory())
 
 
 @bp.route('/location_manage_data', methods=['GET', 'POST'])
@@ -266,10 +347,75 @@ def location_manage_data():
         data = data.items
     data_list = []
     for i in data:
-        dit = {'id': i.id
-            , 'rfid': i.rfid
+        dit = {'rfid': i.rfid
+            , 'num': i.num
             , 'warehouse': i.warehouse
             , 'create_date': str(i.create_date)}
         data_list.append(dit)
     dic = {'code': 0, 'msg': 'SUCCESS', 'count': total, 'data': data_list}
     return dic
+
+
+@bp.route('/get_rfid_data', methods=['GET', 'POST'])
+@login_required
+def get_rfid_data():
+    rfid = request.args.get('rfid')
+    stock = StockModel.query.filter_by(rfid=rfid).first()
+    if not stock:
+        return jsonify({'code': 400, 'msg': 'RFID不存在'})
+    return jsonify({'code': 200, 'rfid': rfid, 'num': stock.num, 'batch': stock.batch, 'kg': stock.kg, 'diameter': stock.diameter, 'lm': stock.lm, 'msg': 'SUCCESS'})
+
+
+@bp.route('/check_rfid_data', methods=['GET', 'POST'])
+@login_required
+def check_rfid_data():
+    rfid = request.args.get('rfid')
+    if check_rfid(rfid):
+        return jsonify({'code': 200, 'msg': 'RFID可用'})
+    else:
+        return jsonify({'code': 400, 'msg': 'RFID已存在'})
+
+
+def check_rfid(rfid):
+    stock = StockModel.query.filter_by(rfid=rfid).first()
+    if stock:
+        return False
+    else:
+        return True
+
+
+@bp.route('/stock_check', methods=['GET', 'POST'])
+@login_required
+def stock_check():
+    return render_template('stock_manage.html', )
+
+
+@bp.route('/stock_check_data', methods=['GET', 'POST'])
+@login_required
+def stock_check_data():
+    page = int(request.args.get('page', 1))
+    limit = int(request.args.get('limit', 50))
+    key = request.args.get('key', '')
+    if key:
+        data = StockModel.query.filter(StockModel.rfid.contains(key)).order_by(
+            db.text('-create_date'))
+        total = data.count()
+    else:
+        data = StockModel.query.order_by(db.text('-create_date')).paginate(page, limit, error_out=False)
+        total = data.total
+        data = data.items
+    data_list = []
+    for i in data:
+        dit = {'id': i.id
+            , 'rfid': i.rfid
+            , 'num': i.num
+            , 'lm': i.lm
+            , 'diameter': i.diameter
+            , 'kg': i.kg
+            , 'batch': i.batch
+            , 'state': i.state
+            , 'create_date': str(i.create_date)}
+        data_list.append(dit)
+    dic = {'code': 0, 'msg': 'SUCCESS', 'count': total, 'data': data_list}
+    return dic
+
